thunar-dropbox-plugin: fixup after #9935
Close #9935.
I don't think it's good to split thunarx into a separate derivation in
this way. We would have the library and associated stuff twice and two
expressions, etc. Distributions do these things by an analogy to multiple-output
derivation (Debian does, for example), so we could use that, but it
doesn't seem important in this case and would be better after #7701 anyway.